requirements.txt
playwright
requests
flask
openpyxl
python-dotenv
